+++ tests/python/unittest/test_numpy_op.py
@@ -173,6 +173,7 @@ def test_np_tensordot():
                 np_out = _np.tensordot(a.asnumpy(), b.asnumpy(), axes)
                 assert_almost_equal(mx_out.asnumpy(), np_out, rtol=1e-3, atol=1e-5)
 
+
 @with_seed()
 @npx.use_np_shape
 def test_np_sum():
